Introduced HostFactory and Parameters
Also switched to JUnit 5

+++ src/main/java/org/openstreetmap/josm/plugins/ods/geotools/GtFeatureSource.java
@@ -21,7 +21,7 @@ public class GtFeatureSource implements OdsFeatureSource {
     private final String featureName;
     private SimpleFeatureType featureType;
     private final String idAttribute;
-    private final long maxFeatures;
+    private final Integer maxFeatures;
     private DataStore dataStore;
     SimpleFeatureSource featureSource;
     CoordinateReferenceSystem crs;
